Edited MrBayesPostProc
Fixed issue with cleanup of temporary files, added t option, improved
section that removes mrbayes block from NEXUS file, updated versioning

+++ MrBayesPostProc/MrBayesPostProc.sh
@@ -4,12 +4,12 @@
 #  __  o  __   __   __  |__   __                                                         #
 # |__) | |  ' (__( |  ) |  ) (__(                                                        # 
 # |                                                                                      #
-#                             MrBayesPostProc v1.3, May 2017                             #
+#                          MrBayesPostProc v1.4, December 2017                           #
 #  SHELL SCRIPT FOR POST-PROCESSING OF MrBayes OUTPUT FILES ON A SUPERCOMPUTING CLUSTER  #
 #  Copyright (c)2017 Justinc C. Bagley, Virginia Commonwealth University, Richmond, VA,  #
 #  USA; Universidade de Brasília, Brasília, DF, Brazil. See README and license on GitHub #
-#  (http://github.com/justincbagley) for further information. Last update: May 8, 2017.  #
-#  For questions, please email jcbagley@vcu.edu.                                         #
+#  (http://github.com/justincbagley) for further information. Last update: December 2,   #
+#  2017.  For questions, please email jcbagley@vcu.edu.                                  #
 ##########################################################################################
 
 ############ SCRIPT OPTIONS
@@ -18,21 +18,24 @@ MY_RELBURNIN_FRAC=0.25
 MY_SS_ANALYSIS_SWITCH=0
 MY_SS_NGEN=250000
 MY_SS_DIAGNFREQ=2500
+MY_TEMP_FILE_SWITCH=1
 
 ############ CREATE USAGE & HELP TEXTS
-Usage="Usage: $(basename "$0") [Help: -h help] [Options: -b s g d] workingDir 
+Usage="Usage: $(basename "$0") [Help: -h help] [Options: -b s g d t] workingDir 
  ## Help:
   -h   help text (also: -help)
 
  ## Options:
   -b   relBurninFrac (def: $MY_RELBURNIN_FRAC) fraction of trees to discard as 'burn-in'
   -s   SS (def: 0, no stepping-stone (SS) analysis conducted; 1, run SS analysis) allows
-       calling stepping-stone analysis starting from NEXUS in current working dir.
+       calling stepping-stone analysis starting from NEXUS in current working dir
   -g   SSnGen (def: $MY_SS_NGEN) if 1 for SS above, allows specifying the number of total 
        SS sampling iterations (uses default number of steps, 50; total iterations will 
        be split over 50 steps) 
   -d   SSDiagFreq (def: $MY_SS_DIAGNFREQ) if 1 for SS above, this specifies the diagnosis 
        (logging) frequency for parameters during SS analysis, in number of generations
+  -t   deleteTemp (def: 1, delete temporary files; 0, do not delete temporary files) calling
+       0 will keep temporary files created during the run for later inspection 
 
  OVERVIEW
  Runs a simple script for post-processing results of a MrBayes v3.2+ (Ronquist et al. 2012)
@@ -59,7 +62,7 @@ if [[ "$1" == "-h" ]] || [[ "$1" == "-help" ]]; then
 fi
 
 ############ PARSE THE OPTIONS
-while getopts 'b:s:g:d:' opt ; do
+while getopts 'b:s:g:d:t:' opt ; do
   case $opt in
 
 ## MrBayesPostProc options:
@@ -67,7 +70,8 @@ while getopts 'b:s:g:d:' opt ; do
     s) MY_SS_ANALYSIS_SWITCH=$OPTARG ;;
     g) MY_SS_NGEN=$OPTARG ;;
     d) MY_SS_DIAGNFREQ=$OPTARG ;;
-
+    t) MY_TEMP_FILE_SWITCH=$OPTARG ;;
+    
 ## Missing and illegal options:
     :) printf "Missing argument for -%s\n" "$OPTARG" >&2
        echo "$Usage" >&2
@@ -92,7 +96,7 @@ echo "$USER_SPEC_PATH "
 
 echo "
 ##########################################################################################
-#                             MrBayesPostProc v1.3, May 2017                             #
+#                          MrBayesPostProc v1.4, December 2017                           #
 ##########################################################################################"
 
 ###### Prep files and then Summarize trees, their posterior probabilities, and their errors using MrBayes.
@@ -102,25 +106,39 @@ echo "INFO      | $(date) | STEP #1: SETUP VARIABLES. "
 	calc () {
 	   	bc -l <<< "$@"
 	}
-    if [[ -f ./*.NEX ]]; then
+	if [[ -s "$(NEX_FILES=./*.NEX; echo $NEX_FILES | head -n1)" ]]; then 
         echo "INFO      | $(date) |          Fixing NEXUS filename... "
     (
         for file in *.NEX; do
             mv "$file" "`basename "$file" .NEX`.nex"
         done
     )
-    fi
+	fi
 
-	MY_NEXUS=./*.nex
+
+	## This script was written to expect only a single NEXUS file in pwd; however, users 
+	## will probably from time to time mistakenly run the script on a directory with 
+	## multiple NEXUS files. Here, we can account for this by taking the first .nex file
+	## found in the directory, by using the following line instead of MY_NEXUS=./*.nex 
+	## (which would only work with 1 file). $MY_MRBAYES_FILENAME is the name of the output
+	## of the run, which will be the root/prefix of each output file. The mb path is self
+	## explanatory.
+	if [[ -s "$(nex_FILES=./*.nex; echo $nex_FILES | head -n1 | sed 's/\ .*//g')" ]]; then 
+		MY_NEXUS="$(ls ./*.nex | head -n1 | sed 's/\ //g')"
+	fi
 	MY_MRBAYES_FILENAME="$(ls | grep -n ".mcmc" | sed -n 's/.*://p' | sed 's/\.mcmc$//g')"
 	MY_SC_MB_PATH="$(grep -n "mb_path" ./mrbayes_post_proc.cfg | awk -F"=" '{print $NF}')"
 
 
 echo "INFO      | $(date) | STEP #2: REMOVE MRBAYES BLOCK FROM NEXUS FILE. "
-	MY_MRBAYES_BLOCK_START="$(grep -n "BEGIN mrbayes;" ./*.nex | sed 's/:.*$//g')"
-	MY_HEADSTOP="$(calc $MY_MRBAYES_BLOCK_START-1)"
-
-	head -n$MY_HEADSTOP $MY_NEXUS > ./simple.nex
+	MY_MRBAYES_BLOCK_START="$(grep -n "BEGIN MrBayes\|Begin MrBayes\|BEGIN mrbayes\|Begin mrbayes\|begin mrbayes" $MY_NEXUS | sed 's/\:.*//; s/\ //g')"
+	if [[ "$MY_MRBAYES_BLOCK_START" -gt "0" ]] || [[ -s "$MY_MRBAYES_BLOCK_START" ]]; then
+		MY_HEADSTOP="$(calc $MY_MRBAYES_BLOCK_START-1)"
+		head -n"$MY_HEADSTOP" "$MY_NEXUS" > simple.nex
+	elif [[ ! "$MY_MRBAYES_BLOCK_START" -gt "0" ]]; then
+		echo "INFO      | $(date) |          NEXUS file contains no MrBayes block. Renaming NEXUS to 'simple.nex'... "
+		mv "$MY_NEXUS" simple.nex
+	fi	
 
 
 echo "INFO      | $(date) | STEP #3: CREATE BATCH FILE TO RUN IN MRBAYES. "
@@ -133,9 +151,9 @@ quit" > ./batch.txt
 
 ##--Flow control. Check to make sure MrBayes batch file was successfully created.
     if [[ -f ./batch.txt ]]; then
-        echo "INFO      | $(date) |          MrBayes batch file ("batch.txt") successfully created. "
+        echo "INFO      | $(date) |          MrBayes batch file ('batch.txt') successfully created. "
     else
-        echo "WARNING!  | $(date) |          Something went wrong. MrBayes batch file ("batch.txt") not created. Exiting... "
+        echo "WARNING!  | $(date) |          Something went wrong. MrBayes batch file ('batch.txt') not created. Exiting... "
         exit
     fi
 
@@ -148,10 +166,6 @@ echo "INFO      | $(date) | STEP #4: SUMMARIZE RUN AND COMPUTE CONSENSUS TREE IN
 
     $MY_SC_MB_PATH  < ./batch.txt > Mrbayes_sumtp_log.txt &		## Use batch to run MrBayes.
 
-
-echo "INFO      | $(date) | STEP #5: CLEANUP FILES. "
-	rm ./batch.txt							## Remove temporary files created above.
-	rm ./simple.nex
 
 
 if [[ "$MY_SS_ANALYSIS_SWITCH" -eq "1" ]]; then
@@ -166,7 +180,15 @@ quit" > ./SS_batch.txt
 
 fi
 
-rm ./SS_batch.txt
+
+echo "INFO      | $(date) | STEP #6: CLEANUP FILES. "
+## If user desires, remove temporary files created above.
+	if [[ "$MY_TEMP_FILE_SWITCH" -eq "1" ]]; then
+		if [[ -f ./batch.txt ]]; then rm ./batch.txt; fi
+		if [[ -f ./simple.nex ]]; then rm ./simple.nex; fi
+		if [[ -f ./SS_batch.txt ]]; then rm ./SS_batch.txt; fi
+	fi
+
 
 echo "INFO      | $(date) | Done with post-processing of MrBayes results using MrBayesPostProc. "
 echo "INFO      | $(date) | Bye.
